refactored some schedule functionalities, started working on DB match implementation
