+++ frontend/src/components/Portals/schedule/scheduleModalUtils.jsx
@@ -1,0 +1,14 @@
+export const weekDays = [0, 1, 2, 3, 4, 5, 6];
+export const timeHour = ["00", "01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23"];
+export const timeMinute = ["00", "05", "10", "15", "20", "25", "30", "35", "40", "45", "50", "55"];
+
+export const colors = [
+    'bg-pink-200 border-pink-400',
+    'bg-rose-200 border-rose-400',
+    'bg-violet-200 border-violet-400',
+    'bg-blue-200 border-blue-400',
+    'bg-green-200 border-green-400',
+    'bg-yellow-200 border-yellow-400',
+    'bg-orange-200 border-orange-400',
+    'bg-stone-200 border-stone-400'
+];
