+++ frontend/src/components/Widgets/optionWidgetConfig.jsx
@@ -1,0 +1,19 @@
+export const optionWidget = () => {
+    /*
+            <div className="text-sm min-w-[90%] max-w-[90%]">
+            <Select
+                styles={{ menuPortal: base => ({ ...base, zIndex: 9999 }) }}
+                menuPortalTarget={document.body} 
+                isSearchable="true"
+                onChange={(e) => handleSelectedContent(e)}
+                options={templateWidget.widget.contents[0].options.map(option => ({ value: templateWidget.id, label: option }))}
+                placeholder={templateWidget.widget.name + "..."}
+                defaultValue={selectedContent[templateWidget.id] !== undefined && { label: selectedContent[templateWidget.id], value: 1 }}
+                getOptionValue={(option) => option.label}
+            />
+        </div>
+    */
+    return (
+        <></>
+    )
+}
