+++ frontend/src/components/Widgets/widgetUtils.jsx
@@ -1,0 +1,5 @@
+export const WidgetType = Object.freeze({
+    MEDIA: "MEDIA",
+    OPTION: "OPTIONS",
+    TEXT: "TEXT"
+});
