+++ frontend/src/components/index.jsx
@@ -9,5 +9,5 @@ export {default as MovableDiv} from "./MovableDiv"
 export {default as MediaModal} from './Portals/MediaModal';
 export {default as PendingMonitorsModal} from './Portals/PendingMonitorsModal';
 export {default as WidgetsModal} from "./Portals/WidgetsModal";
-export {default as ScheduleModal} from './Portals/ScheduleModal';
+export {default as ScheduleModal} from './Portals/schedule/ScheduleModal';
 export {default as FunctionModal} from './Portals/FunctionModal'
